+++ src/Components/Banner.js
@@ -1,14 +1,13 @@
-import logo from '../header.jpeg'
 
 const bannerStyles = {
     width: "100%",
-    height: 200,
+    height: 100,
     objectFit: "none",
     objectPosition: "center",
 }
 
 function Banner() {
-    return <img className="" style={bannerStyles} src={logo} alt="logo" />
+    return <div style={bannerStyles} className="text-center fs-1 text-success text-bg-dark p-3">Friendly Bets</div>
 }
 
 export default Banner;
