+++ src/Components/index.js
@@ -1,7 +1,9 @@
 import Banner from "./Banner";
 import Table from "./Table";
+import Buttons from "./Buttons";
 
 export {
     Banner,
-    Table
+    Table,
+    Buttons
 }
